Moved sklearn classes to separate folder
